extract file addressing to struct
